AssetUtils: Add in getGeolocLocationAtTime for Refill Task TO prepphrase
InventoryPlugin: Add publishRefillTask() which finds the appropriate
maintain inventory tasks and hooks the new Refill to the MI parent and
it's workflow.
Also added an accessor for getting myOrganizations and currentTimeMillis for
used by the RefillGenerator.

RefillGenerator: lots of code to get the Refill task filled in properly.

Rusty: The V scoring function seemed to be the closest match... but its not
exactly what we drew.  Do we need to create a new Scoring Function impl for
ourselves?