ignore level 2 inventory while processing level 6 projections
start on reworking level 2 processing to work with both level 2 demand
coming in from a customer AND level 2 demand that we need to generate
locally from level 6 data.

Need to double check on level 2 processing logic.
Still need to define level 2 asset to help create level 2 inventory if there
is not already one and also to grab the physical pg info.