Change the max for level 6 and 2 to 225 days which is the total time of the oplan - also make this the default value for the operating mode so that in the absence of a sensor or the condition servlet (or the AE) the plugins will process tasks in full level 6 detail for the entire oplan.
Add a check in the RefillProjGenerator to make sure that the end of
the level 2 window is greater then the end of the level 6 window
before processing level 2 tasks.  Note that these boundaries can now
be equal for full level 6 and if level 2 is less than level 6 its a
problem anyway.