add a getLastRefillRequisition method.
Also initialize the inventory levels array with the policie's initial level
for 180 buckets instead of just bucket 0.  Otherwise the refill generators
don't start setting the level until today + OST, so the level will be wrong
after the first bucket until today + OST AND the refill generator thinks
the starting inventory level is 0 - while the allocation assessor thinks
the start is the initial level.  This caused funning inv levels on the charts

Partial fix for bug 1752

If we don't have any actual demand yet then the start bucket to process
project withdraws should be today.  It gets the start bucket from
getfirstprojection in the bg which checks the customer hash for actuals.  If
its empty it was returning the last bucket of the due out list.  Now if its
empty it signals that to the AA by sending back a -1.  If the AA gets a -1
back it will use the today bucket for the first projection bucket.

This fixes missing allocation results on project withdraw tasks that but up
to the end of the due out list.

See bug 1710