If we don't have any actual demand yet then the start bucket to process project withdraws should be today.  It gets the start bucket from getfirstprojection in the bg which checks the customer hash for actuals.  If its empty it was returning the last bucket of the due out list.  Now if its empty it signals that to the AA by sending back a -1.  If the AA gets a -1 back it will use the today bucket for the first projection bucket.
This fixes missing allocation results on project withdraw tasks that but up
to the end of the due out list.

See bug 1710

Use getLastRefillRequisition to tell whether to look at results for
actual refills or projection refills.  This is because when we allocate
project withdraws we may not have any actual demand so it starts processing
at bucket 0. BUT... we could have actual refills so check and use actual
refills instead of projection refills.

Partial fix for bug 1752