head merge from branch for these 3 checkins in the branch
date: 2002/07/21 02:34:18;  author: gvidaver;  state: Exp;  lines: +38 -7
fix for bug #1695 -- spurious tasks show up on added list twice
once in course of running before persistence, and then again after rehydration -
- possibly a core bug
----------------------------
revision 1.46.4.2
date: 2002/07/21 01:24:15;  author: afedyk;  state: Exp;  lines: +49 -2
Bug #1708
InventoryPlugin
  Add two subscriptions for ProjectWithdraw and Withdraw tasks.
Call HandleRemovedRequisitions() and HandleRemovedProjections with the
removed collection of the appropriate subscription.

SupplyExpander
Simplified the remove method to just remove each task in the collection
from the BG.
----------------------------
revision 1.46.4.1
date: 2002/07/18 18:33:29;  author: bdepass;  state: Exp;  lines: +0 -38
Remove not used predicates suppleexpansionpred and projectexpansionpred.
Fix expansion pred (used for both supply and projectsupply) to not check that
the quantity preference is greater than 0 since this will return false for
all projectsupplies since they use rates.  Also - this is an expensive check in
a predicate - we can check for 0 quantities and or rates in other places.

Part of the fix for non allocated projectwithdraws - bug 1710