Bug #1861 The LogisticsInventoryBG maintains a central array of demand. When demand is added the list is updated and likewise, when demand is removed the list is updated.  The code was trying to do an efficient update of the list to up only the elements that are affected by the removed task.  This becomes problematic because the code needs to determine how the task was added, by preference or by EstimatedResult.
Upon rehydration a second level of complexity is added.  The code attempts to remove a task that was never added.

Added a boolean to removeWithdrawProjection() to inform methods dependant upon the demandProjectionArray that the array needs to be rebuilt.  Also added a method to regenerate teh demandProjectionArray when necessary.  The module will still update the demandProjectionArray as tasks are added.