AllocationAsessor: rewrite of main processing methods to contain one main processor of projectwithdraw and withdraw results.  Needs more shortfall testing, but things seem much more deterministic.
InventoryPlugin:
change processing of refill allocations as the new AA compares results
so the processing needed to be of the complete subscription on
backwards flow only - backwardFlow inventories set is created in the
ex alloc and passed to the AA.

ExternalAllocator:
changes to accomodate the reordering of allocation result processing.

SupplyExpander:
cleanup of allocation result processing method

InventoryLevelGenerator:
Fix findCommittedRefill to work for ProjectSupply and Supply tasks
(this method is used by the AA)

LogisticsAllocationResultHelper:
Some changes to accomodate a bug in core in AllocationResult.

LogisticsInventoryBG:
rewrite an update results method to go with the AA changes

RefillGenerator:
Remove a Math.max we had put in to try to fix clips of the critical
level. This fix is not needed with the new AA and it caused the refill
generator to order extra.