Fix bug 2115 - inventory plugin not processing new tasks after rehydration.
Fix is that we are looking at the added and changed lists for the inventory policy - need to look at underlying subscription upon rehydration.

Theory is that SRA used to generate a new policy object upon rehydration so we didn't run into this earlier this month as the code had not changed.