Probable fix for bug 2265.
Tested in tiny - needs to be verified in small and full socieities.

My hunch is that subsistence tasks are being left undisposed because we ignore them depending on their time frame when we fall behind and then we never process the ones we ignored.  This would be a special case for subsistence since subsistence doesn't do any special level 2 processing.  Since level 2 tasks are not generated and because level 6 tasks maybe sent (completed as far as the demand generator is concerned) before we fall behind, the plugin may never wake up again to process the ignored tasks.

At this point I have a fix that for subsistence type inventory plugin instances, a subscription is created for its operating mode so that when the OM changes back to full level 6 processing the plugin processes all inventories to catch unallocated tasks.  This is a little extra work - but its limited by the amount of times we switch back to full level 6 (and it only calls the extra code for subsistence).