2nd half of fix for bug 2265.
First half fixed completion problems with ignored subsistence tasks
after going from level 6 to level 2 and back to level 6.
2nd half fixes shortfalls in subsistence under the same circumstances
as refill projections out towards the end of the oplan are not created
as the need to create them is initially ignored when we are in level 2
processing.  Also add an additional check to make sure the refill
projection generator runs when demand tasks have been changed.

The related external allocator change is to deal with changed demand
tasks that may have already been allocated.