further fix for 2265 all classes.
moved the om change check further up in execute to make sure we always
see them - also add a case to run the AA in downward flow if the om
was changed to get all inventories processed as we may have already
ignored the results.  These 2 cases came up in more testing with the
sensor standin by switching the conditions at different times.