-Change prep phrase building code to only use its own prep phrases. Right now the parent is only giving us the ofType prep and since we create on of these ourselves, this avoids duplicate prep phrases.
-Change the geoloc prep phrase to use the end date(long) of the task
1000 instead of getCurrentTimeMillis() so that the demand tasks go to
whereever we are in theater during the orgact that the task falls in
instead of our home base.  The (-1000) is to avoid a fencepost problem
where the last ProjSupply task's end time can touch the boundary of
the last intheater orgactivity which doesn't seem to be inclusive.
This causes our last PS tasks to be sent to our home base - this
problem was in last year's code as well.