Changes for bug 2779.  Tested at TIC in full on new machines - 20 for 20.
AA:
allocateEarlyProjections on a customer basis so that each one uses
their own lastActualSeen.

ExAll:
remove call to BG.rebuildCustomerHash that wasn't actually ever
getting called anyway.

InventoryPlugin:
Add in calls for rebuildCustomerHash on the touchedBGs when we get new
demand.

LogisticsInvBG and logisticsprops.def
Add a getter for customerHash