Consumption rates are being cached by type id and storing a prototype. Level 2 assets are looked up in the DB by org, but were not being cached by org.
So whichever org on the same node first looked up a
level2BulkPOL or Level2Ammunition consumption rate would have it's
values stored in the cache and subsequent lookups to this typeId
from different orgs would have the same value as the 1st orgs lookup.

This fixes it by having the key be the org and typeId appended as the
combo lookup key.

Tested in a full.  Approved by Beth.