Lots of checks to ensure Task time preferences are reasonable (avoiding WARNs from PiecewiseLinearScoringFunction).
1) daysBetween should not be negative (log at WARN if it is)
2) if best <= early, then don't bother putting early in the scoring
function
--- It's INFO if best==early, otherwise WARN. When RefillProjGen has
best==early, it is often (very frequently) a case of wanting the Task
START to be the given startTime -- usually arrivalInTheater
3) if end <= late, then don't bother putting end in the scoring function
log at WARN when this happens in RefillProjectionGenerator, since I've
not seen it there.
Log at INFO when this happens in GenerateProjectionsExpander. This
happens frequently because it is a Task END_TIME preference and the
BEST is OplanEnd.

The result is many Task ScoringFunctions with only 3 points. It is
more efficient and possibly avoids errors.

On a tiny-trans-UA through stage 1 it means you see no WARNs from
ScoringFunction