Fix cut and paste java doc for changeTask. If prev task and new task are == return null as the javadoc and the calling code expects instead of new task. Change the method signature of changeTask to remove static. Remove static method signatures for comparePreferences and comparePreferencesInner and make comparePreferences private for now since its only called from within.
In comparePreferencesInner only use MoreMath.nearlyEquals with a tolerance of
0.0001 for non START and END time aspects (quantity aspects).  Otherwise
tasks with close timestamps are nearly equal. For example:
prev task with dates of
8/28 - 10/26
and new task with dates of:
8/30 - 10/26
return nearly equal!!! This means the diff-based code doesn't change these!

Make the tolerance for START and END time aspects less than 30 minutes for
now to cover hourly and daily tasks.